+++ test/julian_calendar_test.exs
@@ -1,5 +1,5 @@
 defmodule Cldr.Calendar.Julian.Test do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   test "that we can localize a julian date" do
     import Cldr.Calendar.Sigils