+++ test/week_calendar_test.exs
@@ -1,5 +1,6 @@
 defmodule Cldr.Calendar.Week.Test do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
+
   import Cldr.Calendar.Sigils
 
   defmodule Sunday do
@@ -32,13 +33,13 @@ defmodule Cldr.Calendar.Week.Test do
 
   test "that days of the last month of a long year is 35 or 42" do
     assert Cldr.Calendar.NRF.days_in_month(2012, 12) == 35
-    assert Cldr.Calendar.ISOWeek.days_in_month(2015,12) == 35
+    assert Cldr.Calendar.ISOWeek.days_in_month(2015, 12) == 35
 
     assert Cldr.Calendar.NRF.days_in_month(2013, 12) == 28
-    assert Cldr.Calendar.ISOWeek.days_in_month(2016,12) == 28
+    assert Cldr.Calendar.ISOWeek.days_in_month(2016, 12) == 28
 
-    assert Sunday.days_in_month(2012,12) == 42
-    assert Sunday.days_in_month(2013,12) == 35
+    assert Sunday.days_in_month(2012, 12) == 42
+    assert Sunday.days_in_month(2013, 12) == 35
   end
 
   test "day of week for ISOWeek calendar is correct" do
@@ -116,7 +117,7 @@ defmodule Cldr.Calendar.Week.Test do
 
   test "sunday week calendar" do
     {:ok, today} = Date.new(2018, 53, 3, Sunday)
-    current_period = Cldr.Calendar.month(today)
+    current_period = Cldr.Calendar.Interval.month(today)
     first_day_of_period = current_period.first
     last_day_of_period = current_period.last
     assert first_day_of_period == %Date{calendar: Sunday, day: 1, month: 48, year: 2018}
@@ -126,7 +127,7 @@ defmodule Cldr.Calendar.Week.Test do
   test "that previous month and next month actually are for a week calendar" do
     {:ok, today} = Date.new(2019, 4, 4)
     {:ok, today} = Date.convert(today, Sunday)
-    this_period = Cldr.Calendar.month(today)
+    this_period = Cldr.Calendar.Interval.month(today)
 
     previous = Cldr.Calendar.previous(this_period, :month)
     assert previous.first == %Date{calendar: Sunday, day: 1, month: 44, year: 2018}
@@ -140,7 +141,7 @@ defmodule Cldr.Calendar.Week.Test do
   test "previous month for a week calendar transitioning to prior quarter" do
     {:ok, today} = Date.convert(~D[2019-04-11], Monday)
     prior_period_day = Cldr.Calendar.previous(today, :month, coerce: true)
-    prior_period = Cldr.Calendar.month(prior_period_day)
+    prior_period = Cldr.Calendar.Interval.month(prior_period_day)
 
     last_day_of_prior_period = prior_period.last
     {:ok, gregorian_prior_last} = Date.convert(last_day_of_prior_period, Cldr.Calendar.Gregorian)