+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# Changelog for Cldr Calendars v0.6.0
+
+This is the changelog for Cldr v0.6.0 released on June 1st, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_calendars/tags)
+
+### Breaking Changes
+
+* Moved `year/1`, `quarter/1`, `month/1`, `week/1` and `day/1` to a new module `Cldr.Calendar.Interval`
+
+### Enhancements
+
+* Adds `Cldr.Calendar.Interval.compare/2` to compare two intervals (date ranges) using [Allen's Interval Algebra]((https://en.wikipedia.org/wiki/Allen%27s_interval_algebra) definitions.
+
+* Defaults the calendar to `Cldr.Calendar.Gregorian` for `Cldr.Calendar.year/2`, `Cldr.Calendar.quarter/3`, `Cldr.Calendar.month/3`, `Cldr.Calendar.week/3` and `Cldr.Calendar.day/3`
+
 # Changelog for Cldr Calendars v0.5.0
 
 This is the changelog for Cldr v0.5.0 released on April 21th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_calendars/tags)